moved speakers test files, to tests folder
